Create Reduce the sequence.py
You have an initial numerical sequence: 1, 2, ... , N. At each iteration every second element of the sequence is removed. At the first iteration elements are removed starting with the first element of the sequence. At the second iteration - from the last element of the sequence towards the first element, at the third, again, from the first element, and so on.

0. Suppose N=8 and the sequence is [1 2 3 4 5 6 7 8]
1. The first iteration. Remove every second element starting with the first element: [2 4 6 8]
2. The second iteration. Remove every second element starting with the last element: [2 6]
3. The third iteration. Remove every second element starting with the first element:  [6]
So, for N=8 the answer is 6!